Super stable build (so stable it's fit for a horse). Had several merge conflict issues with Sourcetree, so checked out both branches and manually merged files (both branches had many changes between each other and it didn't want to combine them, just overwrite entire files).
The following is now functional:
- level editor now only loads one oldman and background
- audio now plays correctly
- corrected PlayerController for death fall. Character position now warps upwards on y-axis from "death point" and script invokes Stop() in PlayGame.cs
- other small fixes

The following is still broken:
- wall jumping
- level editor clearing when re-editing twice (I have an idea how to fix this)

+++ Game/Assets/Scripts/ButtonSceneSwitcher.cs
@@ -5,7 +5,7 @@ using UnityEngine.SceneManagement;
 
 public class ButtonSceneSwitcher : MonoBehaviour {
 
-	public int time; // Wait time before scene switch
+	public float time; // Wait time before scene switch
 
 	public void switchScene() {
 		StartCoroutine (executeSwitch ());
